Move blockmachine and lsystem to utils.
These are questionable is they are suppose to part of the server, I will
move them to utils for now.
